- Bug fixes  - Before launch: yaw-stick to left and roll-stick to right = calibrate gyro's (make sure model is still!)  - 1MB dataflash chip supported

+++ Firmware/rtos_pilot/control.h
@@ -45,8 +45,8 @@ struct ControlConfig
 	unsigned int use_pwm : 1;
 
 	enum MixTypes servo_mix;	
-	float max_roll;
-	float max_pitch;
+	double max_roll;
+	double max_pitch;
 	
 	int aileron_differential;  //! Since 0.4
